backend/requirements.txt
# AEON Video Backend - Production Dependencies

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and Security
PyJWT==2.8.0
cryptography==41.0.7
httpx==0.25.2

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
supabase==2.0.2
postgrest==0.13.2

# AI and Video Processing
replicate==0.22.0
openai==1.3.7
anthropic==0.7.8
elevenlabs==0.2.26

# Video Processing
ffmpeg-python==0.2.0
opencv-python==4.8.1.78
Pillow==10.1.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# HTTP and Networking
aiohttp==3.9.1
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
psutil==5.9.6

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
supabase==2.0.2
python-jose[cryptography]==3.3.0
slowapi==0.1.9
python-magic==0.4.27
